Extract main layout
